없는 게시글의 경우 메세지 표시, 게시글 수정시 수정날짜 표시

+++ src/main/java/com/ldh/exam2/jpaBoard/article/controller/ArticleController.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,7 +17,7 @@ public class ArticleController {
     @Autowired
     private ArticleRepository articleRepository;
 
-    @RequestMapping("/list")
+    @RequestMapping("showList")
     @ResponseBody
     public List<Article> showList() {
         return articleRepository.findAll();
@@ -29,7 +30,7 @@ public class ArticleController {
         return article;
     }
 
-    @RequestMapping("detail")
+    @RequestMapping("showDetail")
     @ResponseBody
     public Article showDetail(long id) {
         Optional<Article> article = articleRepository.findById(id);
@@ -38,13 +39,20 @@ public class ArticleController {
 
     @RequestMapping("doDelete")
     @ResponseBody
-    public void doDelete(long id) {
+    public String doDelete(long id) {
+
+        if (articleRepository.existsById(id) == false) {
+            return "%d번 게시글을 찾을 수 없습니다.".formatted(id);
+        }
+
         articleRepository.deleteById(id);
+        return "%d번 게시글이 삭제되었습니다.".formatted(id);
     }
 
     @RequestMapping("doModify")
     @ResponseBody
     public Article doModify(long id, String title, String body) {
+        
         Article article = articleRepository.findById(id).get();
 
         if (title != null) {
@@ -54,6 +62,8 @@ public class ArticleController {
         if (body != null) {
             article.setBody(body);
         }
+
+        article.setUpdateDate(LocalDateTime.now());
 
         articleRepository.save(article);
